refactor(api): extract shared/httpauth ParseBearer + consolidate MCP endpoint-annotation docs (unblock-tv8.55)
pt3: new apps/api/shared/httpauth package; rewire auth/authhandler.go and mcp/mcp.go to httpauth.ParseBearer; move parseBearer subtests to httpauth_test.go.
pt1: transport.go doc anchor aligned with the E1371-elided method annotation.
pt2: consolidate the duplicated E1371-elision paragraphs in mcp.go package doc to a single authoritative annotation note.
